+++ doc/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### added
+- `Segment` class, attrs-based
+
+### changed
+- `Sequence` class is now attrs-based, has factory functions, is itself
+just a list of `Segment`s
+
 ## 0.2.0a1
 ### added
 - add Crowsetta class with simple interface for converting any annotation to