+++ doc/CHANGELOG.md
@@ -4,13 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 0.2.0a2
 ### added
 - `Segment` class, attrs-based
+  + has `asdict` method (wrapper around `attrs` function)
 
 ### changed
 - `Sequence` class is now attrs-based, has factory functions, is itself
 just a list of `Segment`s
+  + now has `to_dict` method
+- `Crowsetta` class is now called `Transcriber` 
 
 ## 0.2.0a1
 ### added